sim: always reset parameters in the LR test
This is related to issue #1746 (it may fix it).

The run method of LikelihoodRatioTest

- fits the null model
- samples the parameters from this
- use the null model best fit as the starting point for a fit to
  the alt model
- grab the parameter values for the alt fit [*]
- runs multiple simulations, based on the samples, where
  - replace null model fit params with a sample (aka proposal)
  - simulate the data using the null model
  - fit the null model (should not change much as we are fitting
    the simulated data starting at this position)
  - fit the alt model
- restore the alt model with the alt fit from [*]

However, in the loop (the multiple simulations part) we do not
reset the alt model after each loop, which can mean that if one
simulation ends up in a "funky" part of the parameter space we
can then end up stuck there in following steps.

There are multiple ways to address this, and I went with a simple
solution that resets both the null and alt models after each
simulation step (the null model should not be a problem within
the model as we explicitly set these values with a new proposal
but it is better to be explicit here, and it also works to make
sure we have reset things after the last simulation).

This makes sense to me, but is statistically sensible?

We can see the difference from the powerlaw / powerlaw + gaussian
test in sherpa/sim/tests/test_sim.py::test_lrt. The ratios
calculated for 25 iterations were (on linux, as they appear to
end up slightly different on macos, maybe partly related to this
issue):

[2.43733721e+00,  2.96321789e+00,  4.18396324e+00,  1.73659109e+00,
 6.61185908e-01,  8.35946480e+00,  1.44437003e+00,  4.21729534e+00,
 1.04513467e+00,  5.68434189e-14, -0.00000000e+00,  1.05796316e+01,
 5.68434189e-14, -0.00000000e+00,  4.26325641e-14,  5.59324922e+00,
 7.10542736e-14,  8.52651283e-14,  3.55271368e-14,  7.10542736e-14,
 5.81173267e+00,  1.24017433e+00,  9.94759830e-14,  2.22883237e+00,
 2.84217094e-14]

We can see that there's a large number of ~0 ratios, which I claim
are examples of where the fit has got "stuck" (in this case I believe
the gaussian fwhm has gone to a ridiculously-large value, something
much larger than 10^5 and hence the model is effectively flat
over the data space). With the hcange in behavior the ratios look,
at least to me, to be more sensible:

[2.43733721, 3.52628288, 4.18396336, 1.73659659, 3.70862308, 2.19009678,
 0.99262327, 2.47817863, 4.29140984, 8.66538795, 0.75845443, 2.19029536,
 1.08725896, 1.58262032, 2.91657742, 0.70781436, 0.79954232, 2.5850919,
 2.0543363,  0.4747516,  8.69094441, 2.35362447, 2.23331886, 4.20676696,
 3.56214367]