sim: store the fitted parameters for the alternate model (ppp)
The parameter values for each fit to the alternate model are
included in the likelihood results as the assumption is that
this is not going to be too much data, and it will help users
identify problem cases - e.g. maybe where the parameter range
needs to be tightened up to avoid getting caught in a "bad" part
of parameter space (although fixing #1746 may make this less
important).

I added a test of issue #1745 so we can note if that is ever
changed (this is, is wstat an allowed statistic for the ppp
test?).

There is also a small change to the debug messages, to use
the "deferred/lazy" string formatting (mainly to lease pylint
as it likes to complain about these lines).