utils: make sure OutOfBoundErr is an actual exception
pylint noted that sherpa.utils.OutOfBoundErr was used as an exception
but did not derive from the exception class. I had thought that it
wasn't actually used, so had left a note to look at this. In working
on a pylint-inspired clean up of sherpa.estmethods I noted that it
is used as an exception (both catching and throwing) so make sure
we derive it from the exception class.