estmethods: pylint clean up
There is a lot of code cleanup needed for this module, and this
commit does not address all the possible changes. This is primarily
because

a. there is currently no test of the code to be changed (mainly
   f-string changes, but also some clean ups to use modern Python
   idioms, use non-Python keywords as arguments/variable names, and
   some use of mutable arguments (mainly {})

b. the fix for the non-Python keyword names would technically be a
   breaking change, and while this is all low-level code, I did not
   want to change too many things at once

The changes made here include

- use str(x) rather than x.__str__()
- use f-strings
- remove or rewrote code that
  - __init__ definitions that just call the super class
  - un-needed __str__/__repr__ calls
    - technically these could be useful, but this is all low-level
      code, the code I have removed does not add much to the default
      behavior, and some of the code was just wrong (e.g. __rep__
      rather than __repr__)
  - broke up 'if/elif/else' chains that return or raise an error
  - clean up class declarations (e.g remove "()" from "class Foo():")
- replaced some use of variable names that over-ride Python keywords,
  such as list, dict, and dir, but many remain
- re-raise an exception from the calling exception to improve the
  traceback experience

I noted several places where we could improve things, in part by
renaming variables, but in part understanding if we even need to
be sending in the value in the first place (as it appears to be
immediately over-written in the routine).

The code is written in an "interesting" style, looking a bit more
like C++ of Java than Python. For instance, the use of classes
local to a class. It is likely not worth changing this at this
time, as we don't have great tests of this functionality.