sim: improve data checks for ReSampleData
In trying to add checks to improve coverage, I realised that the
code - which was meant to report an error if ReSampleData was
called with an invalid data object - did not actualy work, since
it basically required there to be a y attribute, which is not
guaranteed for non-1D data types. So, I've moved the check to
earlier, taking advantage of the ndim field I recently-ish added,
rather than check on the data class (I added an earlier commit
with a test of this functionality).

Should this class warn or error out if given a DataPHA or Data1DInt
object? Alternatively, the code could be enhanced to properly
handle this (but that is hard for the PHA case as we need to worry
about the response).