Remove pkg_resources
The pkg_resources module has long been "somewhat" deprecated but
it looks like, with pip 23.1 and Python 3.12, this is more
important. Fortunately it is an easy fix - see

https://importlib-resources.readthedocs.io/en/latest/migration.html#pkg-resources-resource-string

but we don't need to worry about using the importlib-resources
backport as we are now Python > 3.7. However, it turns out that
the importlib.resources.files interface was only added in Python
3.9, and we still support Python 3.8. So we use importlib_resources
got pre-3.9 systems (I investigated other approaches, including
using the "deprecated" importlib.resources interface) but none are
particularly useful.

Joy.