Tests: AttributeError messages differ widely with Python version
The test_psf_set_model_to_bool check turned out to be fragile,
as the AttibuteError message raised depended on Python version -
so 3.9, 3.10, and 3.11 all had different text. The code tried to
suppport this, but the 3.11 change is too large to make it worth
checking.

This also holds for the astro test.