CI: catch unexpected warning messages
The CI runs of https://github.com/sherpa/sherpa/pull/1752 after
pulling in this branch https://github.com/sherpa/sherpa/pull/1753
started to fail because of a warning message that I could not
replicate. A search lead to this set of issues

- https://github.com/pypa/setuptools/issues/2104
- https://github.com/pypa/setuptools/issues/2052
- https://github.com/pypa/setuptools/issues/1383

which suggests there's a subtle interaction between pip and
setuptools that we are falling over. As we currently have setuptools
pinned to < 60, whilst the distutils changes coming in Python 3.12
work through the ecosystem, the OTS code we run is likely to
hit this issue (i.e. nothing we can fix) so we can just skip the
warnings. Hopefully we can remove these additions soon.