Tests: minor improvements to sim testing
We now use the SherpaVerbosity class for checking the logging
output (and enforve there isn't any) rather than doing it directly
(and not checking).

A test using the old_numpy_printing fixture has been switched to
use the current method, so we can better check what users will
see.

A test has been added to check that the LikelihoodRatioTest errors
out with invalid statistics objects (it tests the existing behavior
so includes a check for WStat as it currently fails, see #1745).