simulate: allow WStat to be used with LikelihoodRatioTest
As well as adding support for the WStat statistic, this also makes
some code-cleanup changes to use lazy-logging - e.g. %s encoding
rather than msg + str(x) - and removing the unused info definition.

The logging instance uses "sherpa" as a name, rather than __name__.
I have added a note to look at this in the future, but for now have
left it as is.