Fix #1762
This is a localized fix for #1762, and addresses the problem by
calling set_coord on the dataset *after* creating the object,
rather than sending the coordinate into the object.

This assumes that the backend "read a FITS image" routine only
ever returns the logical units for the independent axes, which
should be correct (as what else should it return?).

There is a longer-term discussion to be had about how we want to
handle creating an image with "non-logical" coordinates and, if
so, how the transforms between other systems should work, but
that is a sufficiently-larger issue that it is not addressed here.