data: ensure name field is set when validating the DataIMG structure
If the coord argument to DataIMG was invalid - in that it referenced
a coordinate system for which there was no data - then the code would
fall over because there was no name attribute in DataIMG, not because
the argument was invalid.

This is fixed by ensuring the name attribute is set, rather than
leaving it to the superclass.