optmethods: remove mutable argument: OptMethod.fit/NelderMeadxx
The fit method of OptMethod used a mutable argument for
the statkwargs argument (which was set to an empty dictionary).
This has been addressed by making the parameter default to None,
which is then converted to {} in the routine.

Similarly, the NelderMead3 and NelderMead4 __call__ methods
used an argument of finalsimplex=[0, 1, 1], so these now use None
as a short-form for this default.

There are still some methods that use lists - that is ncoresNelderMead
and ncoresNelderMeadRecursive - which I have not changed because I am
not convinced we actually use this code, and although I think I
understand how we could change this, I am not convinced enough.

See #95