optmethods: remove un-needed methods in child classes
The main change is that we no-longer create a method which just
calls the superclass, as we can rely on Python to do this for us
and it makes the code a lot cleaner, as you no-longer have to
check "does this method do something special/different".

A number of pointless return statements have been removed.

TODO comments have been added to point out

- a call that looks like it must fail, because the seed argument
  is not being sent in; I am not willing to try and fix this
  because it really is not clear if we actually use this code
  or not (and a number of follow-up commits are planned to make
  this easier to discover)

- we call the NumPy seterr routine globally when I think it
  should be scoped to this class; this is issue #1238

- point out we import _ncpus but don't use it

- asked if we really need the NelderMead6/6 subclasses

Some commented-out code (importing modules) has been removed,
but not all.