optmethods: code clean up
A few changes, generally suggested by pylint

- remove return call
- f-string changes
- replace else/elif blocks when the previous command returned
  or raised an error

A number of routines create debug-stype information with print
calls. I am not looking at converting this to use the Sherpa
logging infrastructure as that is a big change in its own
right and would derail this work with little noticeable
improvements to the user (relative to planned changes in this
sequence of PRs).