optmethods: switch to argparse from optparse
A lot of the code in these modules could be moved either to
a test file, or a separate script, as it's not useful to the
user of Sherpa. See #1737.

As part of addressing this, move to use the more-modern argparse
module rather than optparse (fortunately it's basically a drop-in
replacement here). The code controls various tests: ideally this
would be part of our actual test suite, or run as additional tests,
but it's too large a change for a single commit.