Tests: add a test of #1779
We want to check both the

    plot_model()

and

    plot_ft()

calls (well, there equivalent), as the PHA model classes are, for
some reason, different, so we need to check both.

We decide to ensure we check

- ungrouped and grouped data
- filtered and unfiltered data

I have created issue #1784 to follow-up on the correct choice of
bin center when grouping values and "normalising by the factor
value.