utils: remove the numpy_convolve routine
I noted a while back (Sep 2018, in #303) that numpy_convolve did
not seem to be used in our code base. This commit removes it. I
note that prior to the comment changes I made in #303 this code
does not appear to have been significantly changed since it was
originally added in the first git commit (April 2015).

It could instead be marked as a deprecation, with removal coming
at a later date, but for such a relatively-simple routine is it
worth it?