multiprocessing: clean up processing of parallel.numcores field
The config parser has long supported fallback values, so we can
make use of this rather than worry about catching a NoSectionError
(which would not have caught the case of a missing or mis-typed
numcores argument). The parser also guarantees that values have
leading and trailing spaces removed so we can remove a strip
call.