parallel: use multi and ncpus rather than _multi and _ncpus
Use the "better" names of multi and ncpus for these two values.
The original names are still accessible, but only from the
sherpa.utils module, and this support is deprecated (but not
warned about other than a comment in the code).

I wanted to make these actual "constants", but it's not really
possible in Python (looks like property needs to be within a
class) and I decided not to stress too much about it. I labelled
them as Final, but this is only relevant for type checking and for
documentation, and does not have any effect at runtime. There's
also the possibility we might want to make these changeable at
run time...