Tests: ensure repeatable test of get_draws
The test_covar_as_none and test_covar_as_argument had the
possibility of failing (seen very occasionally) since they did
not fix the random seed. So the tests have been updated to

- use a fixed seed
- use the same routine for checking the result, to point out
  that the aim of these tests is to show the results are the
  same

Unfortunately, we can't easily check this because

- the fit results, and hence covariance matrices, are different
  when using cash vs cstat (not hugely so, but different enough)
- numerical precision when sending in the covariance matrix vs
  grabbing it from the fit

So we don't actually do much checks of the response. We do add
a check on the accept flags as they appear to be constant for this
seed choice.

There are also two minor code cleanups in this test

- remove a print statement I'd added long ago when debugging something
- switch to using SherpaVerbosity for the module-level hide_logging
  fixture