Fix #1784
For nu fnu plots we need to decide how to calculate the center of each
bin. We had been using code which averaged the centers of groups, but
this is not ideal (although as this is only meaningful for
approximations, it may not matter too strongly) so "do it properly".

Ideally we'd have a nice way to grab this data, but the code would
need a significant re-work to address this, as there are numerous
places we use this logic, all slightly different, and reviewing it is
a large job.