Tests: xspec edge-handling cases
Check some error conditions - eg. using an unknown table name - to
ensure that we catch the error. These are often coming from XSPEC
C++ code so it would be nice if we knew we were creating any extra
string output, but it's not clear how best to do that.

Three unrelated changes

- use the tmp_path fixture for providing access to a temporary
  directly fo a test rather than using routines from the tempfile
  module
- a test of set_xsabund() with a file containing < 30 elements
- remove the use of the clean_astro_ui fixture as not needed