Inline poly.py and tray.py
In real code, one would write this as separate modules,
but inlining it makes is much easier to read
for doctest-plus (I don't in fact know if there is a way to run doctestplus when them needed modules are within
a "literalinclude", which is not part of the main, installed repro).

pytest --doctest-rst docs/model_classes/usermodel.rst passes after this commit
(but we currently don't run rst docs in CI).