Tests: add tests of issue #1802
The issue pointed out by #1802 is how is the regrid logic decided
for a composite model - in particular for the Unary and Binary
OpModel classes. So add tests that check out a variety of
combinations.

In order to catch when the code behavior changes, these tests are
written as regression tests - in that they check we error out
when we expect to - rather than being marked as XFAIL. I have
noticed that, when working on code areas like this where it can
take quite some time to fix the underlying issue, that hiding a
test as XFAIL can make it hard to notice when the behavior changes
(but still fails). This way we will now that something has changed
if the tests start to pass (or, more likely, fail in new and
excitingly-different ways).