regrid: fix #1802
The problem was that when regrid was called on a BinaryOpModel
we just took the first component that had a regrid method, but
if this component was itself a BniaryOpModel - such as in

    mdl1 * mdl2 * mdl3

which turns into

    BinaryOpModel(*, BinaryOpModel(*, mdl1, mdl2), mdl3)

and so the regrid check will be on BinaryOpModel(*, mdl1, mdl2),
which has a regrid call but calls itself, leading to no-end of
recursion.

The solution is to explicitly handle BinaryOpModel cases and
hunt through the "full tree" for the first model that has a
regrid call (and is not a BinaryOpModel). The search is left-biased,
but there's no obvious benefit to any other approach.

Some alternatives include:

- explicitly storing the regrid "class" when creating the
  BinaryOpModel (probably in a _regrid field) so we can just
  identify it when calling regrid()

  this should work, but makes everyone pay for this support, when
  we can just do it when we actually need it

- we could use the ndim field for the model, and just pick
  RegriddableModel1D or RegriddableModel2D directly, which would
  be a lot easier, but it would not be extensible.

One concern is that this probably doesn't handle UnaryOpModel cases
well, but they are currently unsupported.