Attempt to fix #1804
Unfortunately the code uses tuples and lists for handling
multiple datasets, and so it's not obvious when we have them
and when we don't. We could try and do this properly (i.e. ensure
we always use list or tuple) but it's a lot of code to check and
the point of #1804 is to try not to break existing downstream-code,
and this sort of change could do this.