Fix sphinx reference warnings
Sherpa docs used [1], [2], ... for references in many places.
However, references in Sphinx are always global and thus have to be globally unique.
This PR resolves the issue through a combination of :
- inlining URLS instead of listing them in the reference section
- renaming references to unique handles
- replacing links like `[1]_` with simple strings `[1]`, which will not be rendered as link in html, but will be easily understood by human readers who see the numbered entries in "References".

Those changes should be about neutral for people who read the doctring (they don't get rendered links anyway) and do no harm for the html docs overall.
However, they allow us to test the docs for errors on CI.

This is a follow-up to #1519, which started this clean-up in some places.

There is more discussion in #491, which is copied here for the record:
After some further investigation I can confirm what I said in #491 (comment): Every reference in sphinx is global and thus we cannot use labels like [1], [2], because we simply have too many of them. The intention of Sphinx is more to act like LaTeX, where \citep{Bouke2013} will give you the same reference throughout. You could use \citep{1}, but nobody does because the "1" would the same reference throughout the text.

That's not what we do - we use [1] as a throwaway reference in out API and there are dozens of functions wit ha reference called [1], at least a dozen of them refer to ds9 and another 10 or so to the OGIP memo. If we want to keep per-function references, we have to be more create with reference names to makes them unique, e.g. with names as "ds9", "DS9", "ref1", "foot1", etc.
That's also not ideal because it's inconsistent.

However, even if one does that (I made branch where I tried), that still is not enough, because the same docs are often build in tow different locations:

    https://sherpa.readthedocs.io/en/4.14.1/ui/api/sherpa.astro.ui.calc_kcorr.html
    https://sherpa.readthedocs.io/en/4.14.1/ui/api/sherpa.astro.ui.utils.Session.html#sherpa.astro.ui.utils.Session.calc_kcorr

Are two different web-pages, which are build from the same docstring - once as a method on the Session object and once as a function of sherpa.astro.ui. Because they are build from the same docstring, they necessarily have the same reference name 0 ans thus still trigger the following warning:

/Users/guenther/code/sherpa/sherpa/astro/ui/utils.py:docstring of sherpa.astro.ui.utils.Session.calc_kcorr:69: WARNING: duplicate citation k2, other instance in /Users/guenther/code/sherpa/docs/ui/api/sherpa.astro.ui.calc_kcorr.rst