Resolve Sphinx "WARNING: duplicate object description of ..."
These warnings/errors come from different sources:
- The sherpa.astro.xspec module docstring was included twice
- a number of classes had an attribute manually listed in the docstring, but also had it in the code itself

The goal of this commit is to make sphinx pass with no errors, not to make perfect docstring. In particular, the plot_prefs system needs to be improved and will be touched in upcoming plotting work anyway, so here I'm just making spinx pass clean.