Make Readthedocs fail on sphinx warnings
RTD is part of our CI; failures in the doc build are handily reported on PRs.
With #1808, #1806, #1795, and others, I attempt to put more responsibility on the docs build - it's actually generating figures etc. that will be shown.
So, it makes only sense to also have a strong CI check for that.

This commit will turn mere sphinx warnings into errors and make the build fail; the rationale is that even sphinx warning should be fixed as some of them indicate severe problems, such as a plot not being generated or API links that don't get correctly set.