Replace bing/end/exception with a context manager
Plotting backends now act as context managers. Previously, interactive plots where wrapped into a
```
backend.begin()
try:
    code here
    backend.end()
except:
   backend.exception()
```
This cna be written much more concisely as a context manager.
Changes are small, because none of the existing backends actually does anything
in the begin.end/exception functions, which makes is even more convenient to simply leave the definition of the context manager to the base class.