Tests: improve coverage of sherpa.utils.random
It's not obvious why these are not already tested (or, at least,
why choice is not partly covered), but it may in part be due to
some code only used when multiple cores are in use. As these are
simple checks add them. The tested functions are simple wrappers
around NumPy routines so we do not do extensive testing.