Tests: allow string comparison similar to doctestplus
We have a number of places where we check multi-line string output,
normally for str output (e.g. for fit results), but in a few other
places. To make this easy to handle, introduce a fixture we can use
(check_str) which returns a function with the calling syntax of "(got,
expecteds)". In this case got is the multi-line response and expecteds
is a list of string or re.Pattern objects. For the strings, we use a
simplification of the doctestplus logic used to handle the "# doctest:
+FLOAT_CMP" checks to make it easier (hopefully) to write the tests
(i.e. users don't need to change the expected text, or hopefully not
too much).

I had originally made pytest-doctestplus a requirement, so we could
use the doctest logic rather than an approximation of it, but this
is currently problematic (e.g. see issue #1818) and is a bit of an
invasive change, so I have moved to this approach.