sherpa_test: tweak how pytest/plugins are installed
We had been doing

    try:
        importlib.import_module("pytest!=1.0,!=5.2.3")
    except ImportError:
        go-and-install-pytest()

which ends up always trying to install pytest (which would
be a no-op when already installed). So, this code makes the
installation a bit-more verbose, by separating out the package
name from the installation-constraint.

It was found when trying to add pytest-doctestplus as a
requirement. In doing so I was seeing some bizarre behavior,
which made me add the "check" option (another option for
specifying a dependency). It isn't actually used here (since
this PR does not need any plugin), but I have left the code
in.

I added typing rules for the code changed (but did not want
to add to more of this module as it turns out it would require
some code tweaks).