tests: astro plots internal clean up
The PHA source-plot tests used a "non-standard" approach to defining
the energy response (using the bin_lo/hi field of the DataPHA
structure rather than using a RMF). Switch to a RMF to make it
clearer what is going on.

It also sets the PHA exposure time to 10 seconds, which does not
change the tests now, but will once issue #1825 is addressed.