PHA source plots now handle rate setting and fix factor setting
Source plots of PHA data - so

  plot_source
  plot_source_component
  the background variants

did not follow the set_analysis rate setting (ie they always plotted
"per second" values). This is now addressed (but is unlikely to be
very useful).

The plots also did not handle the factor setting correctly when not
at the default value of factor=0. For factor=1 and 2 it would
multiply by the bin width (factor=1) or bin width^2 (factor=2)
rather than the bin value (mid-point of the independent axis).
This check also adds support for factor>2 (to match the model/data
behavior), even though this is not expected to be used.

Fix #1825