model: improve examples in docstrings
Allow model.py, parameter.py, and regrid.py to be included in
the docstring pytest run.

This involved

- tweaking a couple of examples to actually pass (since the
  environment they run in is a bit different from how they
  were created)

- change the choice of variables just to avoid having to create
  new objects just for the test

- hide some tests where it's just not worth creating objects
  used in the test (e.g. too complex to create the necessary
  state and it would hide from the actual example).

Note that one test in regrid.py is technically not-useful but
this has already been noted as issue #840 so the example is
kept (but we now note the warning message that is created), as
this PR is just about testing the existing code.