astropy: rework RMF input (fix #1830)
Rework the RMF input code so that instead of flattening columns
containing variable-length rows immediately, we instead do this
as part of the final "flattening" process. It was hard for me
to understand the old code so I felt this was a cleaner approach,
and hopefully we have enough tests to check the various code
paths.

The old code allowed some columns - in particular MATRIX - to
be optional. We now more-closely follow the OGIP standard and
make the columns required.

The tests now pass (so have had the XFAIL marker removed) when
run with the pyfits backend.