xspec: updates to parameters made in XSPEC 12.13.1
Update several parameters from XSPEC 12.13.1. These include changing
the default value or the minimum value (some of these are from a
report we made to HEASARC about the parameter minimum leading to
invalid results, thanks to a user ticket). Changes incoude

  - cemkl, cevmkl
  - bwcycl
  - gabs (just changes the units)

Note that the bwcycl parameter delta has been renamed to del in
XSPEC but we do not make this change since del is likely to get
confused with the del() routine.