XSPEC: update test for XSPEC 12.13.1
We reported an issue with how XSPEC table models are set up from
the "model interface", and how it was different to the set up done
from XSPEC (the program). This appears to have been fixed in
12.13.1 (although not documented). Amusingly enough, we were
accidentally relynig on the broken behavior in a single test:
we evaluated a table model outside the desired range and so rather
than returning 1, as it should have, it was returning -1. Well,
it now returns 1.