Tests: check we fail as expected
The check_str fixture needs to be sent an array of strings as
it's second argument otherwise it will do a character-by-character
comparison, which is not what we want here. So

- split tests up so if one fails the others are still run
- switch to an array for the expected text
- use the match argument to pytest.raises to make sure we are
  failing for the expected reason

This means that the tests will need to be updated if the
pytest.approx assertion error ever changes, but let's worry
about that once it happens.