I/O: Allow RMFs to be written out
This requires

- an "entry point" in sherpa.astro.io
- support in the backends
- plumbing in the sherpa.astro.ui layer

This is significantly-more work than the ARF case because we have
to

a) reconstruct the RMF on-disk data given the flattened version
   stored in DataRMF
b) have to decide how best to store the F_CHAN, N_CHAN, and MATRIX
   blocks (basically, do we use a VARIABLE LENGTH FIELD)

The reconstruction is relatively easy, it is just a bit involved:

- need to handle rows where N_GRP = 0
- for rows where N_GRP > 1 need to combine multiple input rows
  to combine the F_CHAN, N_CHAN, and MATRIX data
  - we force Int32, Int32 and Float32 types for these
    as the downstream processing can convert them to Int16 if wanted

Since there's no obvious way to create a variable-length field - that
is, crates and AstroPy have different ways of handling and creating
such data - we create a ndarray with dtype=object for columns we want
stored as a VLF, and expect the backend to do the conversion (for
crates we don't have to do anything; there's a little-bit of work
for AstroPy).

For deciding whether a column should be VLF or not we use a few simple
rules (which do not match those used by the data producers, including
CIAO):

- if N_GRP is a single value then not a VLF
- if N_GRP < 4 for all rows then create 2D arrays with the Y axis
  being max(N_GRP)
  - this is based on the OGIP documentation that claims it is generally
    faster and less memory-intensive to do this
    https://heasarc.gsfc.nasa.gov/docs/heasarc/caldb/docs/memos/cal_gen_92_002/cal_gen_92_002.html

We need to write out a TLMIN4 keyword for RMF data (to list the
minimum channel value) but crates does not appear to have a way
to do this at this time, soI have written a little bit of code
using the cxcdm module to do this. cxcdm is part of CIAO and if
you have crates then you will have cxcdm!