I/O: update pyfits backend to better handle output types
This is a partial update for the AstroPy backend (pyfits_backend.py)
to use the Table interface (and converting to FITS HDUS when necessary)
rather than creating FITS objects directly.

This should provide a more-consistent interface with the choice of
output - that is, ASCII or FITS - and it fixes an existing problem
that the old code did not write out FITS columns with the correct
types (e.g. np.float32 would get written out as np.float64 and the
integer types would also not have their types preserved). This was
because the FITS columns were being created with

    format=<dtype converted to upper case>

rather than the FITS format specifiers, such as "I", "J", "E", and
"D".

There are some other subtle changes introduced in this commit - mainly
that the header of ASCII files now is

    # <name1> ...

rather than

    #<name1> ...

I should really look at #1091 to see what that does.