ui: add get_plot_prefs/get_contour_prefs calls
For users who want to change the delchi plot preferences this
replaces

    prefs = get_resid_plot(...).plot_prefs or .histo_prefs

with

    prefs = get_plot_prefs("resid")

It's a small use-of-life improvement, and we could decide to
deprecate the get_data/model_plot/contour_prefs() calls.

There is a subtle issue with the "fit" option, since get_fit_plot()
returns

    .dataplot
    .modelplot
    .plot_prefs   which is just {}

and the code would (by defualt) return the plot_prefs setting,
which does not do anything. I decided to make "fit" error out
rather than have it potentially confuse users. Since the plot
option also has to worry about "bkg_fit" from the astro layer
we can not just check on "fit".

Another issue with the plot command is we allow **kwargs to
be set: this means we can easily support the "bkg_id" argument
for the bkg-style plots in the astro layer without having to
overload get_plot_prefs. The downside is that the documentation
is harder to read, as we do not have a bkg_id keyword.

It's essentially the same thing for the contour plots, but simpler,
as:

- do not have to deal with plot_prefs vs histo_prefs
- do not have to worry about extra arguments from "bkg" commands
  in the astro layer