plots: fix #1828
The problem in #1828 turned out to be because xlo was an integer
array and xhi was a floating-point array. In this case the
combination, done by sherpa.plot.utils.intersperse(), created an
integer array, rather than a floating-point one.

For most users this situation is unlikely to occur (since the
independent axis will be sent in via a Sherpa data object, in
some form, and so the change in data types is unlikely). However,
we can easily fix this by changing from using intersperse() to
using numpy.vstack((xlo, xhi)).T.flatten(), as this handles the
necessary type conversions.

The intersperse routine has not been changed, other than to
add a note about data-type compatability. It is likely we can
mark this as deprecated, but I don't have the energy to do this
analysis right now.